requirements.txt
tensorflow>=2.21.0
keras>=3.12.1
numpy>=2.2.6
flask>=2.3.0
numpy>=1.23.0
matplotlib>=3.10.0
seaborn>=0.13.0
scikit-learn>=1.7.0
pillow>=8.0.0
werkzeug>=2.3.0
